requirements.txt
contourpy==1.1.1
cycler==0.11.0
fonttools==4.42.1
importlib-resources==6.1.0
kiwisolver==1.4.5
matplotlib==3.7.3
numpy==1.24.4
packaging==23.1
Pillow==10.0.1
pyparsing==3.1.1
python-dateutil==2.8.2
six==1.16.0
zipp==3.17.0
